chore: clean gameStateMachine extraction per thermo review
- Trim dispatcher imports to delegates only (~195 lines)
- Remove duplicate type exports and dead code from action handlers
- Import BindingContractTerm normally; export SECTOR_IDS from shared board config
