+++ packages/shared/src/index.ts
@@ -65,6 +65,7 @@ export {
   PERIMETER_SIZE,
   PERIMETER_TILES,
   SECTOR_HUB_POSITIONS,
+  SECTOR_IDS,
   SECTORS,
   TOTAL_BOARD_MARKET_VALUE,
   UTILITY_POSITIONS,
