+++ packages/web/package.json
@@ -10,6 +10,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@oligopoly/shared": "workspace:*",
     "@oligopoly/validation": "workspace:*",
     "@simplewebauthn/browser": "^13.3.0",
     "react": "^19.2.4",
