+++ packages/web/src/lib/sectorIds.ts
@@ -1,13 +1,1 @@
-/** Canonical sector ids (matches @oligopoly/shared board config). */
-export const SECTOR_IDS = [
-  "emerging_tech",
-  "big_tech",
-  "finance",
-  "healthcare",
-  "energy",
-  "defense_media",
-  "elite_tech",
-  "fast_track",
-] as const;
-
-export type SectorId = (typeof SECTOR_IDS)[number];
+export { SECTOR_IDS, type SectorId } from "@oligopoly/shared";
